fix ManualEnter UI
